Contains all changes for experiments on speech
DualPurposeSpeechService contains some code over riding the existing
speech recognizer. NewTaskActivity contains code capturing raw audio
data. This part actually works. StartLiveCardServiceAtBootReceiver
contains script to make sure the live card starts on install or after
booting the device. This works.  ToDoLiveCardActivity contains the
changes to override the VoiceListener. This does not work as this class
is never used. We use a passive LiveCard so binding of LiveCard happens
in the class ToDoLiveCardService class. Need to find a way to check this
code by assigning it to a proper Activity. NewTaskActivity might be a
good idea.